HRDS_v01

+++ HRDS.Web/Models/Entities/EducationGrade.cs
@@ -5,7 +5,7 @@ namespace HRDS.Web.Models.Entities;
 
 public partial class EducationGrade
 {
-    public int GradeId { get; set; }
+    public decimal GradeId { get; set; }
 
     public string GradeCode { get; set; } = null!;
